requirements.txt
google-genai
python-dotenv
pyttsx3
edge-tts asyncio pygame